rename & flatten apache httpclient package
