+++ apm-agent-plugins/apm-httpclient-core/src/test/java/co/elastic/apm/agent/http/client/AbstractHttpClientInstrumentationTest.java
@@ -22,7 +22,7 @@
  * under the License.
  * #L%
  */
-package co.elastic.apm.agent.httpclient;
+package co.elastic.apm.agent.http.client;
 
 import co.elastic.apm.agent.AbstractInstrumentationTest;
 import co.elastic.apm.agent.impl.TextHeaderMapAccessor;
