+++ apm-agent-plugins/apm-jdk-httpclient-plugin/src/test/java/co/elastic/apm/agent/apachehttpclient/HttpClientInstrumentationTest.java
@@ -22,8 +22,9 @@
  * under the License.
  * #L%
  */
-package co.elastic.apm.agent.httpclient;
+package co.elastic.apm.agent.apachehttpclient;
 
+import co.elastic.apm.agent.http.client.AbstractHttpClientInstrumentationTest;
 import org.junit.Before;
 
 import java.net.URI;
@@ -31,12 +32,14 @@ import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
 
-public class HttpClientAsyncInstrumentationTest extends AbstractHttpClientInstrumentationTest {
+public class HttpClientInstrumentationTest extends AbstractHttpClientInstrumentationTest {
     private HttpClient client;
 
     @Before
     public void setUp() {
-        client = HttpClient.newBuilder().followRedirects(HttpClient.Redirect.NORMAL).build();
+        client = HttpClient.newBuilder()
+            .followRedirects(HttpClient.Redirect.NORMAL)
+            .build();
     }
 
     @Override
@@ -44,11 +47,18 @@ public class HttpClientAsyncInstrumentationTest extends AbstractHttpClientInstru
         HttpRequest request = HttpRequest.newBuilder()
             .uri(URI.create(path))
             .build();
-        client.sendAsync(request, HttpResponse.BodyHandlers.ofString()).get();
+
+        client.send(request, HttpResponse.BodyHandlers.ofString());
     }
 
     @Override
     protected boolean isErrorOnCircularRedirectSupported() {
+        // skip circular redirect test
+        //
+        // this http client just gives up after a fixed amount of redirects
+        // this value can be set with the 'jdk.httpclient.redirects.retrylimit' and defaults to 5.
+        // there is no exception thrown, the response provided to the user is just a redirect response
         return false;
     }
+
 }
